daily updates
want to work on some android projects, want to learn some widely used frameworks, as well as get familiar and comfortable with kotlin language and get comfortable when referring kotlin based github projects. May not update this repository for a short while. Will come back and update when there is a need. Waiting for my 32GB ram.